move deserialize_string_or_vec to util::serde
